client-api: rename `get_state_events_for_key` to `get_state_event_for_key`
since only one event is ever returned on this api
