+++ crates/ruma-client-api/src/state/get_state_event_for_key.rs
